--- src/Senparc.Weixin.Open/Senparc.Weixin.Open/XPayAPIs/XPayJson/CreateFundsBillJsonResult.cs
@@ -1,20 +1,0 @@
-﻿using Senparc.Weixin.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Senparc.Weixin.Open.XPayAPIs.XPayJson
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public class CreateFundsBillJsonResult : WxJsonResult
-    {
-        /// <summary>
-        /// 充值单 id
-        /// </summary>
-        public string bill_id { get; set; }
-    }
-}
