--- src/Senparc.Weixin.Open/Senparc.Weixin.Open/XPayAPIs/XPayJson/DownloadBillJsonResult.cs
@@ -1,20 +1,0 @@
-﻿using Senparc.Weixin.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Senparc.Weixin.Open.XPayAPIs.XPayJson
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public class DownloadBillJsonResult : WxJsonResult
-    {
-        /// <summary>
-        /// 下载地址
-        /// </summary>
-        public string url { get; set; }
-    }
-}
