--- src/Senparc.Weixin.Open/Senparc.Weixin.Open/XPayAPIs/XPayJson/QueryPublishGoodsRequestData.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Senparc.Weixin.Open.XPayAPIs.XPayJson
-{
-    /// <summary>
-    /// 查询批量发布道具任务
-    /// </summary>
-    public class QueryPublishGoodsRequestData
-    {
-        /// <summary>
-        /// 0-正式环境 1-沙箱环境
-        /// </summary>
-        public int env { get; set; }
-    }
-}
