+++ src/Senparc.Weixin.WxOpen/src/Senparc.Weixin.WxOpen/Senparc.Weixin.WxOpen/AdvancedAPIs/XPay/XPayJson/DownloadBillJsonResult.cs
@@ -1,0 +1,20 @@
+﻿using Senparc.Weixin.Entities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Senparc.Weixin.WxOpen.AdvancedAPIs.XPay
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class DownloadBillJsonResult : WxJsonResult
+    {
+        /// <summary>
+        /// 下载地址
+        /// </summary>
+        public string url { get; set; }
+    }
+}
