+++ src/Senparc.Weixin.WxOpen/src/Senparc.Weixin.WxOpen/Senparc.Weixin.WxOpen/AdvancedAPIs/XPay/XPayJson/QueryUploadGoodsRequestData.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Senparc.Weixin.WxOpen.AdvancedAPIs.XPay
+{
+    /// <summary>
+    /// 查询批量上传道具任务
+    /// </summary>
+    public class QueryUploadGoodsRequestData
+    {
+        /// <summary>
+        /// 0-正式环境 1-沙箱环境
+        /// </summary>
+        public int env { get; set; }
+    }
+}
